Add files via upload
The aim of the project is to provide the management of MovieOnRent with actionable insights and data-driven recommendations for stocking their inventory. 
By understanding the most frequently rented movies, popular genres, and actors in demand, the management can make informed decisions about which movies to stock in their stores. 
This, in turn, aims to improve customer satisfaction by offering a wide selection of movies that align with customer preferences. 
Also ,  enhance business profitability by optimizing inventory management and increasing movie rentals.